Fix #150: Rename Resource to ResourceBase and leave deprecation note
